+++ activerecord/test/cases/token_for_test.rb
@@ -3,6 +3,8 @@
 require "cases/helper"
 require "models/matey"
 require "models/user"
+require "models/cpk"
+require "models/room"
 require "active_support/message_verifier"
 
 class TokenForTest < ActiveRecord::TestCase