+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Allow to save records with polymorphic join tables that have inverse of specified.
+
+    *Markus Doits*
+
 *   Fix association scopes applying on the incorrect join when using a polymorphic `has_many through:`.
 
     *Joshua Young*