+++ activerecord/CHANGELOG.md
@@ -2,6 +2,14 @@
 
     *Goulven Champenois*
 
+*   Allow to save records with polymorphic join tables that have inverse of specified.
+
+    *Markus Doits*
+
+*   Fix association scopes applying on the incorrect join when using a polymorphic `has_many through:`.
+
+    *Joshua Young*
+
 *   Allow `ActiveRecord::Base#pluck` to accept hash arguments with symbol and string values.
 
     ```ruby