+++ activerecord/CHANGELOG.md
@@ -1,3 +1,11 @@
+*   Allow to save records with polymorphic join tables that have inverse of specified.
+
+    *Markus Doits*
+
+*   Fix association scopes applying on the incorrect join when using a polymorphic `has_many through:`.
+
+    *Joshua Young*
+
 *   Allow `ActiveRecord::Base#pluck` to accept hash arguments with symbol and string values.
 
     ```ruby