+++ activerecord/CHANGELOG.md
@@ -32,6 +32,11 @@
     *Ruy Rocha*
 
 
+## Rails 7.1.4.1 (October 15, 2024) ##
+
+*   No changes.
+
+
 ## Rails 7.1.4 (August 22, 2024) ##
 
 *   Allow to eager load nested nil associations.